backend/requirements.txt
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
celery==5.3.6
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.23.5
PyJWT==2.8.0
redis==5.0.4
sniffio==1.3.1
sqlparse==0.5.0
torch==2.1.2+cu121
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
